requirements.txt
# appnope==0.1.3
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# asttokens==2.0.5
# attrs==21.4.0
# backcall==0.2.0
# backports.cached-property==1.0.1
# beautifulsoup4==4.11.1
# bleach==5.0.0
# certifi==2022.5.18.1
# cffi==1.15.0
# charset-normalizer==2.0.12
# click==8.1.3
# debugpy==1.6.0
# decorator==5.1.1
# defusedxml==0.7.1
# entrypoints==0.4
# exceptiongroup==1.1.0
# executing==0.8.3
# fastjsonschema==2.15.3
# #Flask==2.2.3
bcrypt==4.0.1
Flask==2.0.3
Flask-Cors==3.0.10
# grpcio==1.46.3
# grpcio-tools==1.46.3
# idna==3.3
# iniconfig==2.0.0
# ipykernel==6.13.0
# ipython==8.3.0
# ipython-genutils==0.2.0
# ipywidgets==7.7.0
# itsdangerous==2.1.2
# jedi==0.18.1
# Jinja2==3.1.2
# jsonschema==4.5.1
# MarkupSafe==2.1.1
# matplotlib-inline==0.1.3
# mistune==0.8.4
# # mysql==0.0.3
mysql-connector-python==8.0.32
# mysql-connector-python==8.0.26
# #mysqlclient==2.1.1
# nbclient==0.6.3
# nbconvert==6.5.0
# nbformat==5.4.0
# nest-asyncio==1.5.5
# notebook==6.4.11
# overrides==6.1.0
# packaging==21.3
# pandocfilters==1.5.0
# parso==0.8.3
# pexpect==4.8.0
# pickleshare==0.7.5
# pluggy==1.0.0
# prometheus-client==0.14.1
# prompt-toolkit==3.0.29
# protobuf==3.20.1
# # psutil==5.9.1
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pycparser==2.21
# Pygments==2.12.0
# pyparsing==3.0.9
# # pyrsistent==0.18.1
# pytest==7.2.1
# python-dateutil==2.8.2
# pyzmq==23.0.0
# qtconsole==5.3.0
# QtPy==2.1.0
# requests==2.27.1
# Send2Trash==1.8.0
# six==1.16.0
# soupsieve==2.3.2.post1
# stack-data==0.2.0
# terminado==0.15.0
# tinycss2==1.1.1
# tomli==2.0.1
# tornado==6.1
# traitlets==5.2.1.post0
# typing-utils==0.1.0
# typing_extensions==4.2.0
# urllib3==1.26.9
# wcwidth==0.2.5
# webencodings==0.5.1
# Werkzeug==2.2.3
# widgetsnbextension==3.6.0
# wrapt==1.14.1
